catlearn-27-wandb converting pytorch derived models to use named_parameters instead of parameters
