+++ catlearn/categorical_model.py
@@ -12,7 +12,7 @@ import torch
 from torch.optim import Optimizer
 
 from catlearn.tensor_utils import (
-    DEFAULT_EPSILON, Tsor, remap_subproba)
+    DEFAULT_EPSILON, Tsor, remap_subproba, AbstractModel)
 from catlearn.graph_utils import DirectedGraph, NodeType
 from catlearn.composition_graph import CompositeArrow, ArrowType
 from catlearn.relation_cache import RelationCache
@@ -22,18 +22,6 @@ from catlearn.algebra_models import Algebra
 # Abstract type definitions
 # NB: docstring temporarily disabled
 # pylint: disable=missing-docstring
-class AbstractModel:
-    @property
-    def parameters(self) -> Callable[[], Iterable[Any]]:
-        raise NotImplementedError()
-
-    def freeze(self) -> None:
-        raise NotImplementedError()
-
-    def unfreeze(self) -> None:
-        raise NotImplementedError()
-
-
 class RelationModel(AbstractModel):
     def __call__(self,
                  src: Tsor,
@@ -47,6 +35,7 @@ class ScoringModel(AbstractModel):
                  dst: Tsor,
                  rel: Tsor) -> Tsor:
         raise NotImplementedError()
+
 
 # pylint: enable=missing-docstring
 
@@ -188,7 +177,7 @@ class DecisionCatModel:
 
 
 
-class TrainableDecisionCatModel(DecisionCatModel):
+class TrainableDecisionCatModel(DecisionCatModel, AbstractModel):
     """
     Specialization of DecisionCatModel working with trainable
     torch modules.
@@ -242,13 +231,12 @@ class TrainableDecisionCatModel(DecisionCatModel):
         """
         return self.algebra.flatdim
 
-    @property
-    def parameters(self) -> Callable[[], Iterable[Any]]:
+    def named_parameters(self) -> Iterable[str, Any]:
         """
         returns an iterator over parameters of the model
         """
-        return lambda: chain(self._relation_model.parameters(),
-                             self._scoring_model.parameters())
+        return chain(self._relation_model.named_parameters(),
+            self._scoring_model.named_parameters())
 
     def freeze(self) -> None:
         """
